Censor the URL as printed
I got tired of having to censor what I paste into Slack, so do the sane
thing and censor the auth tokens when printing them.  This should also
make it safe to run cart verbosely in CI.

I saw that the censoring changed the order of the fields in the URL,
sorting them, and if we care about the URL at all, then something has
gone wrong and having the order between what is logged and what is sent
be wrong is asking for trouble.  So normalize _all_ URLs to "exactly
what the order would be if censored".